# influxdata: add stats gathering