… influxdata/stat_string.go