This is not intended as part of the mainline branch, but
will be used to gather stats about line-protocol use in production,
so it's targetted at another feature branch.

For the record, these are the changes to the benchmarks from adding
the stats:

```
name                                                         old time/op    new time/op    delta
ScanEntriesSkipping/long-lines-8                               16.7ms ± 2%    26.7ms ± 1%  +59.90%  (p=0.008 n=5+5)
ScanEntriesSkipping/long-lines-with-escapes-8                  17.8ms ± 1%    26.0ms ± 1%  +46.73%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-8                         237ns ±12%     271ns ±14%     ~     (p=0.206 n=5+5)
ScanEntriesSkipping/single-short-line-with-escapes-8            228ns ± 4%     315ns ± 1%  +38.28%  (p=0.008 n=5+5)
ScanEntriesSkipping/many-short-lines-8                          130ms ±18%     158ms ± 2%  +21.48%  (p=0.008 n=5+5)
ScanEntriesSkipping/field-key-escape-not-escapable-8            278ns ±14%     282ns ± 1%     ~     (p=0.659 n=5+5)
ScanEntriesSkipping/tag-value-triple-escape-space-8             278ns ± 6%     344ns ± 4%  +23.64%  (p=0.008 n=5+5)
ScanEntriesSkipping/procstat-8                                 2.26µs ± 1%    3.09µs ± 1%  +36.94%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-8                        18.4ms ± 1%    29.3ms ± 1%  +59.10%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-with-escapes-8           93.3ms ± 1%   104.2ms ± 0%  +11.73%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-8                  304ns ± 3%     401ns ±14%  +32.04%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-with-escapes-8     323ns ± 1%     444ns ± 4%  +37.67%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/many-short-lines-8                   153ms ± 1%     216ms ± 0%  +40.77%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/field-key-escape-not-escapable-8     282ns ± 3%     386ns ± 2%  +37.02%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/tag-value-triple-escape-space-8      379ns ±10%     476ns ± 3%  +25.78%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/procstat-8                          5.23µs ± 2%    6.12µs ± 1%  +17.09%  (p=0.008 n=5+5)
Tokenize-8                                                     20.1ms ± 3%    25.4ms ± 3%  +26.89%  (p=0.008 n=5+5)

name                                                         old speed      new speed      delta
ScanEntriesSkipping/long-lines-8                             1.57GB/s ± 2%  0.98GB/s ± 1%  -37.46%  (p=0.008 n=5+5)
ScanEntriesSkipping/long-lines-with-escapes-8                1.48GB/s ± 1%  1.01GB/s ± 1%  -31.85%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-8                       123MB/s ±11%   108MB/s ±13%     ~     (p=0.222 n=5+5)
ScanEntriesSkipping/single-short-line-with-escapes-8          141MB/s ± 4%   102MB/s ± 1%  -27.78%  (p=0.008 n=5+5)
ScanEntriesSkipping/many-short-lines-8                        203MB/s ±16%   166MB/s ± 2%  -18.37%  (p=0.008 n=5+5)
ScanEntriesSkipping/field-key-escape-not-escapable-8          120MB/s ±15%   117MB/s ± 1%     ~     (p=0.690 n=5+5)
ScanEntriesSkipping/tag-value-triple-escape-space-8           180MB/s ± 6%   145MB/s ± 4%  -19.09%  (p=0.008 n=5+5)
ScanEntriesSkipping/procstat-8                                588MB/s ± 1%   430MB/s ± 1%  -26.97%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-8                      1.42GB/s ± 1%  0.89GB/s ± 1%  -37.15%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-with-escapes-8          281MB/s ± 1%   252MB/s ± 0%  -10.50%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-8               95.4MB/s ± 3%  72.7MB/s ±16%  -23.81%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-with-escapes-8  99.2MB/s ± 1%  72.1MB/s ± 4%  -27.33%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/many-short-lines-8                 171MB/s ± 1%   121MB/s ± 0%  -28.96%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/field-key-escape-not-escapable-8   117MB/s ± 3%    85MB/s ± 2%  -27.02%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/tag-value-triple-escape-space-8    133MB/s ±11%   105MB/s ± 3%  -20.97%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/procstat-8                         254MB/s ± 2%   217MB/s ± 1%  -14.61%  (p=0.008 n=5+5)
Tokenize-8                                                   1.31GB/s ± 3%  1.03GB/s ± 3%  -21.18%  (p=0.008 n=5+5)

name                                                         old alloc/op   new alloc/op   delta
ScanEntriesSkipping/long-lines-8                                 512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/long-lines-with-escapes-8                    512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/single-short-line-8                          512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/single-short-line-with-escapes-8             512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/many-short-lines-8                           512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/field-key-escape-not-escapable-8             512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/tag-value-triple-escape-space-8              512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesSkipping/procstat-8                                   512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/long-lines-8                          512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/long-lines-with-escapes-8           19.5kB ± 0%    19.5kB ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/single-short-line-8                   512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/single-short-line-with-escapes-8      512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/many-short-lines-8                    512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/field-key-escape-not-escapable-8      512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/tag-value-triple-escape-space-8       512B ± 0%      512B ± 0%     ~     (all equal)
ScanEntriesWithoutSkipping/procstat-8                            512B ± 0%      512B ± 0%     ~     (all equal)
Tokenize-8                                                       512B ± 0%      512B ± 0%     ~     (all equal)
```